requirements.txt
clickclick>=0.5
PyYAML
keyring
requests
